requirements.txt
Django>=5.0.10,<5.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
psycopg[binary]>=3.1.0
